+++ src/usb/src/lib.rs
@@ -1,4 +1,5 @@
 pub mod command_line;
+pub mod file;
 pub mod monitor;
 pub mod on_off;
 pub mod process_status_code;