+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 ### Added
 - Help option.
+- Export mount info to file.
 
 ### Changed
 - Improve connect/disconnect steps.